Do less work when creating Components in glimmer
We starting noticing that there are quite a bit of work we do in
Components init that are not needed in Glimmer, so we tried to make a
slimmer version. When we eventually get to feature-flagging, we can
conditionally export the right class. We will probably need to add more
stuff as we go (or perhaps we can remove more work), but this is a good
start.